html/semantics/embedded-content/the-model-element/attributes/global/draggable.tentative.html

<!DOCTYPE html>
<html>
  <head>
    <link
      rel="help"
      href="https://html.spec.whatwg.org/multipage/embedded-content-other.html#the-model-element"
    />
    <title>Test for the functionality of the draggable attribute</title>

    <style>
      model,
      img {
        background-color: rgb(255, 0, 0);
      }
    </style>
  </head>

  <body>
    <model draggable="true" width="200" height="200">
      <source src="../../resources/chromeball.usdz" type="model/vnd.usdz+zip" />
      <source src="../../resources/chromeball.glb" type="model/gltf-binary" />
    </model>
  </body>
  <!-- Include the test harness -->

  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>

  <script>
    const model = document.querySelector("model");

    promise_test(async (t) => {
      const modelRect = model.getBoundingClientRect();
      const initialStyle = getComputedStyle(model);
      const dragStartPromise = new Promise(resolve => { 
          model.addEventListener("dragstart", (dragEvent) => {
            model.style.backgroundColor = "rgb(0, 255, 0)";
            resolve();
          });
      });
      
    const dragEndPromise = new Promise(resolve => {
      model.addEventListener("dragend", () => {
        model.style.backgroundColor = "rgb(0, 0, 255)";
        resolve();
      });
     }); 
      assert_equals(
        initialStyle.getPropertyValue("background-color"),
        "rgb(255, 0, 0)",
        "model has valid initial bgcolor"
      );

      await new test_driver.Actions()
        .pointerMove(modelRect.x + 10, modelRect.y + 10)
        .pointerDown()
        .addTick()
        .pointerMove(modelRect.x + 100, modelRect.y + 10)
        .send();
      await dragStartPromise;
      const nextStyle = getComputedStyle(model);
      assert_equals(
        nextStyle.getPropertyValue("background-color"),
        "rgb(0, 255, 0)",
        "model changes bgcolor on drag-move"
      );

      await new test_driver.Actions().pointerUp().addTick().send();
      await dragEndPromise;
      const finalStyle = getComputedStyle(model);
      assert_equals(
        finalStye.getPropertyValue("background-color"),
        "rgb(0, 0, 255)",
        "model changes bgcolor on drag-end"
      );
    }, "Model responds to draggable correctly");
  </script>
</html>
